+++ packages/twenty-front/src/modules/action-menu/actions/types/ActionAvailableOn.ts
